(html) fixed dumbass mistake.
the mistake was me trying to put both information banks next to each other and every time came out awkward.
my solution was to create two divs called left-side and right-side
and go from there. making sure that i floated them both to
opposite sides so i was then able to inline both of them as if they were regular divs (they are btw)

i believeee this update also fixes a few things with the formatting and scaling
as the scaling now isn't awkward as hell and can actually be scaled to different devices
although i havent tried it on my iphone yet.